+++ src/gfx/font_rendering/font.hpp
@@ -29,6 +29,7 @@ struct Font {
   hb::unique_ptr<hb_buffer_t> hbBuffer;
 
   std::string path;
+  std::string familyName;
   float height;     // font size
   float width;
   float dpiScale;
@@ -50,6 +51,8 @@ struct Font {
 
   // OpenType font features (e.g., stylistic sets, character variants)
   std::vector<hb_feature_t> features;
+
+  void SetFeatures(std::string_view featuresStr);
 
   static std::expected<Font, std::runtime_error>
   FromName(const FontDescriptorWithName& desc, float dpiScale);